Rename cyclic.rs -> uniformcyclic.rs
