+++ ndhistogram/src/axis/mod.rs
@@ -6,8 +6,8 @@
 mod bininterval;
 pub use bininterval::bininterval::BinInterval;
 pub use bininterval::singlevaluebininterval::SingleValueBinInterval;
-mod cyclic;
-pub use cyclic::UniformCyclic;
+mod uniformcyclic;
+pub use uniformcyclic::UniformCyclic;
 mod uniform;
 pub use uniform::Uniform;
 mod uniformnoflow;
